fixed detailpage bug
