+++ components/TripList.js
@@ -3,6 +3,7 @@ import Link from "next/link";
 import styled from "styled-components";
 
 export default function TripList({ campingTrips }) {
+  console.log(campingTrips);
   return (
     <>
       <Header>
