+++ pages/detailpage/[id].js
@@ -4,9 +4,10 @@ import DetailCard from "@/components/DetailCard";
 export default function DetailPage({ campingTrips }) {
   const router = useRouter();
   const { id } = router.query;
-
-  let currentTrip = campingTrips.find((campingTrip) => campingTrip.id === id);
-  console.log(campingTrips);
+  if (!id) {
+    return;
+  }
+  const currentTrip = campingTrips.find((campingTrip) => campingTrip.id === id);
 
   return (
     <>
