+++ pages/index.js
@@ -1,14 +1,6 @@
-import { useEffect, useState } from "react";
 import TripList from "@/components/TripList";
 
-export default function ListPage() {
-  const [campingTrips, setCampingTrips] = useState([]);
-
-  useEffect(() => {
-    const storedTrips = JSON.parse(localStorage.getItem("campingTrips")) || [];
-    setCampingTrips(storedTrips);
-  }, []);
-
+export default function ListPage({ campingTrips }) {
   return (
     <>
       <TripList campingTrips={campingTrips} />
